Sized first with the three-target emission (`-stdlib runtime -platforms windows/amd64,linux/amd64,darwin/amd64 -platform-stage`, seeded per the ritual with the seeded .cs count asserted, classified by the tool's own sentinel mtime so only WRITTEN files count): 33/31/33 written-and-differing files per target, ~52 unique, in exactly the families predicted -- a single-target run had reported the linux folder as "identical" because it never wrote it (the L3 rule read from the other side). Positive control 2 passed to the file: linux/package_info.cs's <ImportInitializers> block absent at HEAD and carrying five hooks fresh (internal/abi, internal/bytealg, internal/godebugs, internal/stringslite, runtime/internal/sys), linux/arena.cs still holding the fifth inline in its pre-relocation home. Then four commits, each built on all three GoTargetOS values with 0 errors: (1) c7e8487d1 the init-hook relocation -- four shared files (alg, error, metrics, panic) and all three flavours' arena.cs lose their per-file [GoInit] blocks, all three package_info.cs gain the same five-hook block, and the two tracked-but-never-overlay-refreshed review siblings mfinal.cs.auto / runtime2.cs.auto refreshed byte-for-byte (a regen is exactly where they move); metrics.cs's hook cut here while its d.compute statement stayed train 5's; (2) aec1c8887 the explicit-cast drift, 44 files (19 flat + 9/7/9 per flavour), full-file copies each verified FIRST to carry no remaining diff outside this family -- the hunk rule satisfied by measurement -- every sampled instance an explicit numeric cast around an expression already of that type, representation not value; (3) 3d242f66c the position maps and lift records, the three package_info.cs overlaid whole after the same single-family check, hashes shifting where commits 1-2 touched content, linux/darwin each gaining ~30 GoDynamicTypeLift entries windows already had (missing registrations, not new types); (4) 4df231e5a the go2cs.SynthesizedStructs friend grant in runtime.csproj, traced (independently by R) to converter commit 3f2e02bc0's corpus-wide emission, banked in two csproj until this regen -- regen debt of the A3 grant family, not R's StructOf arc, not orphaned.

Controls at the tip: a dry-run merge against the post-train-5 master resolved metrics.cs cleanly (hook removal and the recvʗ1 hoist in non-overlapping regions, both present); the fresh-emission control the ritual owes -- go2cs.exe rebuilt from 21f7c9677, the same three-target emission against the MERGED tree -- closed 33/31/33 to 1/1/2 line-ending-only artifacts of the verification script comparing an unwritten seed copy against itself, zero real differences with --strip-trailing-cr on all three targets: the merged tree is what the current converter emits, in full; CNR NO REGRESSION byte-identical over 691 packages. Union battery at the merge result on the i7: converter suite, full CNR, runtime.csproj on all three targets, the runtime and reflect -tests builds through the pipeline path.

+++ src/core/runtime/darwin/netpoll_kqueue.cs
@@ -115,7 +115,7 @@ retry:
         // Ignore the ETIMEDOUT error for now, but try to dive deep and
         // figure out what really happened with n == ETIMEOUT,
         // see https://go.dev/issue/59679 for details.
-        if (n != -_EINTR && n != -_ETIMEDOUT) {
+        if (n != (int32)(-_EINTR) && n != (int32)(-_ETIMEDOUT)) {
             println((@string)"runtime: kevent on fd"u8, kq, (@string)"failed with"u8, -n);
             @throw(runtimeNetpollFailedˢ);
         }
